requirements.txt
accelerate==1.4.0
gradio==5.18.0
gradio-client==1.7.2
numpy==2.1.1
Pillow==11.1.0
torch==2.6.0
torchvision==0.21.0
ultralytics==8.3.80